moved Database.kt into a utils package
