+++ app/src/main/java/com/hich2000/tagcapella/tags/TagDTO.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.runtime.toMutableStateList
-import com.hich2000.tagcapella.Database
+import com.hich2000.tagcapella.utils.Database
 import com.hich2000.tagcapella.music_player.SongDTO
 import javax.inject.Inject
