+++ app/src/main/java/com/hich2000/tagcapella/tags/TagRepository.kt
@@ -1,7 +1,7 @@
 package com.hich2000.tagcapella.tags
 
 import com.hich200.tagcapella.TagcapellaDb
-import com.hich2000.tagcapella.Database
+import com.hich2000.tagcapella.utils.Database
 import com.hich2000.tagcapella.music_player.SongDTO
 import javax.inject.Inject
 import javax.inject.Singleton
